更新私聊用推送（IOS）

+++ message-server/src/main/java/com/blt/talk/message/server/manager/RouterHandlerManager.java
@@ -207,7 +207,7 @@ public class RouterHandlerManager {
         PduAttachData attachData = new PduAttachData(usersStatRsp.getAttachData());
 
         if (attachData.getType() == AttachType.HANDLE) {
-            int handleId = attachData.getHandle();
+            Long handleId = attachData.getHandle();
             ChannelHandlerContext msgCtx =
                     ClientUserManager.getConnByHandle(usersStatRsp.getUserId(), handleId);
             if (msgCtx != null) {
